requirements.txt
python-telegram-bot==13.1
Requests